Untested tomcat shutdown and check routine.

+++ egg/egg-tomcat-stop.sh
@@ -26,11 +26,40 @@ if [ "$ISTOMCATSTOPLOCK" == "0"  ]; then
     ./log.sh "Waiting 5 seconds for $APPDIR Tomcat to shut down before sending hard pkill"
     sleep 5
     #Syntax of next line sensitive
-    uselessjibberishvar=`</dev/null ssh -n $HOST "pkill -f egg/${APPDIR}/tomcat/conf; echo terminated"`
+    #uselessjibberishvar=`</dev/null ssh -n $HOST "pkill -f egg/${APPDIR}/tomcat/conf; echo terminated"`
     #Syntax of above line sensitive
-    ./log.sh "uselessjibberishvar=$uselessjibberishvar"
-    ./log.sh "Waiting 40 seconds for $APPDIR Tomcat to fully shut down"
-    sleep 40
+    #./log.sh "uselessjibberishvar=$uselessjibberishvar"
+    #./log.sh "Waiting 40 seconds for $APPDIR Tomcat to fully shut down"
+    #sleep 40
+
+
+    export tcdone="false"
+    while [ $tcdone == "false" ]
+    do
+        tcprocesschk=`ssh $HOST "[ -n \"\\\`pgrep egg/${APPDIR}/tomcat\\\`\" ] && echo 1"`
+        ./log.sh tcprocesschk=$tcprocesschk
+        if [ "$tcprocesschk" != 1 ]; then
+            ALLISWELL=0
+            ./log.sh "Tomcat ${APPDIR} process still running, sending pkill and waiting 5 sec"
+            uselessjibberishvar=`</dev/null ssh -n $HOST "pkill -f egg/${APPDIR}/tomcat/conf; echo terminated"`
+            ./log.sh "uselessjibberishvar=$uselessjibberishvar"
+            sleep 5
+        else
+            ./log.sh "Tomcat ${APPDIR} process not found"
+            export tcdone="true"
+        fi
+    done
+
+
+
+
+
+
+
+
+
+
+
 fi
